fixes for report already exist

+++ app/controllers/reportController.js
@@ -5,30 +5,42 @@ const { allReportWorker } = require('../workers/reports/reportWorker');
 
 const getSingleIssueReport = async (req, res, next) => {
     const response = await singleIssueReport(req.params.path, false, req.params.year)
-    response.pdfResponse.pdf.pipe(response.pdfResponse.output)
-    response.pdfResponse.pdf.on('error', err => {
-        console.error('generateLatexPDF::', err)
-        return res.send(err)
-    })
-    response.pdfResponse.pdf.on('finish', () => {
-        console.log('PDF generated!')
-        fs.unlinkSync(response.pdfResponse.reportPrepd)
+    if (response.pdfResponse) {
+        response.pdfResponse.pdf.pipe(response.pdfResponse.output)
+        response.pdfResponse.pdf.on('error', err => {
+            console.error('generateLatexPDF::', err)
+            return res.send(err)
+        })
+        response.pdfResponse.pdf.on('finish', () => {
+            console.log('PDF generated!')
+            fs.unlinkSync(response.pdfResponse.reportPrepd)
+            return res.send({ report: `${getAppRoute()}/issueReports/${response.reportFile}` })
+        })
+    } else if (response.reportFile) {
         return res.send({ report: `${getAppRoute()}/issueReports/${response.reportFile}` })
-    })
+    }
+
+    return res.send(response)
 }
 
 const getMultiIssuesReport = async (req, res, next) => {
     const response = await multiIssuesReport(req.params.path, false, req.params.year)
-    response.pdfResponse.pdf.pipe(response.pdfResponse.output)
-    response.pdfResponse.pdf.on('error', err => {
-        console.error('generateLatexPDF::', err)
-        return res.send(err)
-    })
-    response.pdfResponse.pdf.on('finish', () => {
-        console.log('PDF generated!')
-        fs.unlinkSync(response.pdfResponse.reportPrepd)
+    if (response.pdfResponse) {
+        response.pdfResponse.pdf.pipe(response.pdfResponse.output)
+        response.pdfResponse.pdf.on('error', err => {
+            console.error('generateLatexPDF::', err)
+            return res.send(err)
+        })
+        response.pdfResponse.pdf.on('finish', () => {
+            console.log('PDF generated!')
+            fs.unlinkSync(response.pdfResponse.reportPrepd)
+            return res.send({ report: `${getAppRoute()}/pathReports/${response.reportFile}` })
+        })
+    } else if (response.reportFile) {
         return res.send({ report: `${getAppRoute()}/pathReports/${response.reportFile}` })
-    })
+    }
+
+    return res.send(response)
 }
 
 const getNationReport = async (req, res, next) => {
